Use the express-annotations rewrite and add support for multiple names
This depends on the rewrite of express-annotations.

Routes can now be mapped to multiple names. When mapping routes to names
the first name the route's path is mapped to becomes its canonical
`name` annotation, but it also maintains an array of all its names on
its `names` annotation.

The `getRouteMap()` method will now return entries for a route under all
of its `names`. A route's canonical `name` is used to resolved
conflicts; if there's another route already mapped to its `name`, then
none of its `names` will be mapped and the route will be skipped.